It's alive!  The system now calls to the server and asks for the list of videos to show on each of the pages.  At the moment it shows the whole list, no discrimination, but that is a battle for another day. Moved some code into appropriate helper classes and had to put some things into async tasks so Android stopped complaining. about having network tasks on the main thread.
By the way, I edited the build.xml in the android sdk directory (that does not show
up in this commit) so that all deprecations warnings show as errors to fix.  I edited
that to add -Xlint:deprecation - see below:

  <!-- compilation options -->
    <property name="java.encoding" value="UTF-8" />
    <property name="java.target" value="1.5" />
    <property name="java.source" value="1.5" />
    <property name="java.compilerargs" value="-Xlint:deprecation" />
    <property name="java.compiler.classpath" value="" />

TODO: clean up imports in all the files, lots o' mess there.